Move staticmethods to crypten
Summary:
Moved `MPCTensor.cat()` and `MPCTensor.stack` to be `crypten.cat()` and `crypten.stack()` to align more closely with pytorch conventions.

Note: There are still dependencies in TTP and Beaver on ArithmeticSharedTensor.stack and BinarySharedTensor.stack in order to vectorize scatters and reveals. I'll work on removing these in the future.

Reviewed By: shubho, vreis

Differential Revision: D16923998

fbshipit-source-id: e0aa555bbb7e638f0609932a472ba24f35c74c2c