Replace scatter calls with shared RNG
Summary:
Implemented PRZS (pseudo-random sharing of zero), using shared RNG. This allows us to remove the scatter from every __init__ call in ArithmeticSharedTensor and BinarySharedTensor.

The intention of this diff was to remove the need for util.sharing, though there are several places where these are still used (wrap_rng, etc.). It is still good to have since it reduces runtime.

Reviewed By: shubho

Differential Revision: D16928362

fbshipit-source-id: 4b22d85696bd989df9897c3f38c0e00ba51b5121