Refactored and updated implementation of ONNX Gather
Summary:
Refactored the ONNX Gather Module in nn/modules.py so that private variables are no longer accessed.

Updated the implementation of Gather to use fancy indexing instead of numpy.take() The implementation currently throws an assertion failure if the dimension (also called axis) is not 0 or 1.

The definition of numpy.take in terms of fancy indexing is given here:
https://docs.scipy.org/doc/numpy/reference/generated/numpy.take.html
However, I am not fully certain that ONNX Gather operation is equivalent to the numpy.take. The two examples in the ONNX docs use numpy.take, but they only use a 1-d index.

Laurens suggested running inference on a ONNX graph with just a Gather operation to see if it matches the definition of  numpy.take (and therefore the encrypted Gather operation). I used the Caffe2 backend to run inference on the ONNX graph, but it only implements axis = 0 or axis = 1. For these two axes, it seems like the behavior of numpy.take matches the ONNX Gather. I did not add this test to the commit, as it seemed orthogonal to the purpose of test_nn. We could still try other backends for ONNX (onnxruntime or tensorflow) but I would have to install them somewhere.

I am also worried that I am missing something. On the AI Platform Users group, they said they implemented ONNX gather in Caffe2 with Caffe2 gather for axis 0 and Caffe2 batch gather for axis 1. If ONNX Gather could be easily implemented by fancy indexing, they wouldn’t need to do something more complex, right?

UPDATE: I verified by using onnxruntime backend on my mac that the equivalence holds for all dimensions (of sample tensors...)

Reviewed By: shubho, knottb

Differential Revision: D16925402

fbshipit-source-id: b49791a93c205fe64244cb93477c810752c3d06b