Renaming onnx_gather to take: Updated
Summary:
Renamed the onnx_gather function to take. Since torch.take() does not allow for a dimension argument, (crypten) take matches torch.take() when the dimension=None, and matches ONNX Gather() when dimension is not None.

Also deleted the previous test_take function in test_mpc.py. The new test_take function covers the old test_onnx_gather function, and also the functionality when the dimension is not set.

FYI, I had to uncommit the previous commit and submit a fresh commit because I was adding back an encrypted_tensor.py file otherwise...

Reviewed By: knottb

Differential Revision: D17074083

fbshipit-source-id: 5f6e39162be5aa69c9d2c9fbcea1af4fedd19ff0