Implement max / min with argmax / argmin outputs
Summary:
Using pytorch `torch.max()` and `torch.min()` functions return just max / min values when no `dim` argument is supplied, but they return `max, argmax` and `min, argmin` when a `dim` argument is supplied.

This diff makes that change. We still return a one-hot output with the same size as the input for our argmax. This is explained in the new docblock. I have also included a TODO to implement a version of argmax/min that gives index-valued outputs rather than one-hot, but I have not implemented it in this diff.

This diff also allows us to supply max / min with `keepdim` arguments as in PyTorch (that also work only when a `dim` argument is given).

This diff also returns the indices of the "argmax_pool2d" (as in the feature request) as the 2nd output of max_pool2d. The MaxPool2d module has been adjusted to account for this. Eventually we should add `return_indices` to the MaxPool2d module (as in `torch.nn.MaxPool2d`) for proper backpropagation.

Finally, the tests have been adjusted to account for `(max, argmax)` tuple outputs and keepdim.

This diff is a first step to implementing an encrypted version of `torch.nn.MaxUnpool2d` as listed here:

https://pytorch.org/docs/stable/nn.html#maxunpool2d

Reviewed By: shubho

Differential Revision: D17170404

fbshipit-source-id: 0f54dce78fca2b9dd174b77d24c2faeec563b7cb