Add autograd
Summary:
This implements an autograd tensor that wraps `CrypTensor`, and implements the corresponding `AutogradFunction`s for nearly all common functions.

The behavior for the autograd is similar to that of PyTorch: the `AutogradFunction`s use the same API, and gradients are only stored in leaves.

The diff implements a registration mechanism for `AutogradFunction`s: this allows the user to define an `AutogradFunction` in their own code, register it under their desired name in the autograd, and then call that function name on the `AutogradCrypTensor`. This allows the user to, for instance, define their own loss functions.

I am still considering merging `AutogradCrypTensor` directly into `CrypTensor`. I think it is possible to do that but there are a few caveats: in particular, one needs to make sure that no tracing of function calls is happening during the backward pass. I will try this in a separate diff.

Reviewed By: shubho

Differential Revision: D17150856

fbshipit-source-id: e1201140a3a2ef1ae7416ad31746c3c36805b9b4