Make modules work on CrypTensors
Summary:
The `crypten.nn` package expects only `AutogradCrypTensor`s as input. This may lead to unexpected behavior for the user when that user specifies another `CrypTensor` (for example, an `MPCTensor`) as input.

This diff resolves that problem by wrapping `Module.forward()` calls such that `CrypTensor` inputs are automatically converted to `AutogradCrypTensor`s before the actual forward call is performed.

The diff adapts the unit tests to test both input types.

Reviewed By: shubho, xingz9

Differential Revision: D17517702

fbshipit-source-id: 39019c186e6b68a0b32f10ed7b2355d49db90747