Fix module test/eval and examples
Summary:
Batchnorm forward was broken during inference. This was because calls to `model.train()` and `model.eval()` did not get called recursively, so child modules were untouched. This diff corrects this.

Additionally, to test this, I ran examples on my local machine. These were broken because some the launchers were not correct, and `crypten.init()` was not called by several of them.

Reviewed By: lvdmaaten

Differential Revision: D17555222

fbshipit-source-id: 4bed97bc5911829eb06cb54cbc901ce905f3d16d