In process communicator (#53)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/53

Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/7

Implements a communicator that does message passing between threads instead of
communicating between multiple processes. This is useful for multiple reasons:
 1) Threads are easier to debug;
 2) This works on Windows (torch.distributed doesn't);
 3) This runs much faster (no spawn/forking involved);

This adds a new base class for our tests (MultiThreadedTestCase) which is
analogous to MultiProcessTestCase, and makes all tests run on both
environments.

To get this working I needed a few seemingly unrelated changes:
 - Implement isend/irecv in DistributedCommunicator; We were previously
   calling torch.distributed directly;
 - Move the PRZS code to crypten.init instead of doing it in the communicator;
   That allows sharing the implementation between multiple communicators;

Reviewed By: lvdmaaten

Differential Revision: D17113829

fbshipit-source-id: db30b12f83c0bc50a92fe57fefef524fb3aa0459