Implement loading for modules in crypten.load (#64)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/64

`crypten.load` assumed that the file contained a torch tensor and would break when the file contained a torch module.

This diff allows the loaded file to contain a module, but requires that the `dummy_model` kwarg contain a module with the same architecture as the loaded module. `crypten.load` will check that the module architectures are the same.

Reviewed By: shubho

Differential Revision: D17634998

fbshipit-source-id: e3aedd78fd4312b985d25f52ce90d4097a995359