Use 2 iterations for log to improve precision (#75)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/75

From the dogfooding:

`Tutorial 2: in the example, seems the difference between public and private method is kind of high, e.g. log(1.5) the public result is 0.4055 while private version is 0.4598. Is that expected?`

We discussed the tradeoffs between accuracy and performance. Person dogfooding was happy to find out we have knobs for this after I suggested using more iterations in the notebook. I think iterations=2 may be a good default here. It is useful to show good accuracy, especially in the tutorials, so I think we should just take the perf hit (extra exp() call and ~3 multiplies). Log doesn’t seem to be a bottleneck anyway for most things.

Reviewed By: shubho

Differential Revision: D17666057

fbshipit-source-id: 3212693b6091ffad8ea1d1478384813054b2f713