Use torch.multiprocessing.spawn in multi process launcher
The existing multi process launcher joins processes in the same order
that they are started. This means that an exception or SIGSEGV goes
unnoticed until process 0 decides to exit (with some error about rank
1 being gone).

Instead, torch.multiprocessing.spawn joins processes in whatever order
they terminate. If they all terminate with exit status 0, this is
considered success. If one of them terminates with a non-zero exit
status, this is considered an error, and the remaining processes are
actively terminated.