Make sure that functions are exposed in CrypTensor (#14)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/14

A major issue in CrypTen is that code that builds on top of `CrypTensor` (such as `gradients.py` and `crypten.nn`) makes assumptions about the existence of functions that are not specified in `CrypTensor` at all.

Unless a function is specified in `CrypTensor`, code on top of it should not be using it.

This diff attempts to correct these problems by attempting to complete the API specification of `CrypTensor`. For some functions of which the implementation is not at all specific to `MPCTensor`, I even moved the implementation of the function into `CrypTensor` to encourage code sharing between different implementations of the `CrypTensor` abstract class.

In the process, documentation is also moved from `MPCTensor` (which is not even included in the documentation) to `CrypTensor` (which is where it belongs).

Please note that this diff also makes it clear that the documentation is still far from complete: many user-facing functions are hardly documented at all.

Reviewed By: shubho

Differential Revision: D17780300

fbshipit-source-id: e433a6089c46014523b46d393257f9961311190f