Parallelize multi-bit B2A converter
Summary:
Vectorized B2A calls in B2A converter in order to parallelize rounds. We skipped this optimization in earlier versions because it was not relevant to our code runs.

I did not change or add to the tests here because our tests cover bits = 1 and bits = 64. All other values for bits do not have nice analogs to check against floating point values. I think this is fine.

Differential Revision: D18246140

fbshipit-source-id: ad41f5b4a355d156dba3ed262712bbaa8cc247d2