Allow interchangeable encoders
Summary:
This diff is in response to https://github.com/facebookresearch/CrypTen/issues/57

It is unclear whether all of the API choices here are good. We had always planned on making encoders modular like this, but given that a lot of the arithmetic code assumes fixed point with scaling, it is unclear whether other encoders can be supported using the same API.

Additionally tests seem to run somewhat slower, but this is likely intermittent.

Differential Revision: D20008428

fbshipit-source-id: d7c2931fa5f29d9048961525fbce505d1325205a