Merge AutogradCrypTensor into CrypTensor (#63)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/63

This diff merges `AutogradCrypTensor` into `CrypTensor`.

As part of this change, the diff contains several breaking changes. In particular, the `forward()` signature of functions such as `sum`, `mean`, `min`, and `max` now matches that of PyTorch exactly. The function signatures in `MPCTensor` also match those of the `forward()` functions for the most part -- although there are still some todos there. Primitives such as `ArithmeticSharedTensor` and `BinarySharedTensor` are no longer `CrypTensor`s because we never want to autograd through them. The functions `crypten.cat` and `crypten.stack` are now proper static functions.

I left a series of todos in the code that need to be resolved in subsequent diffs.

Reviewed By: knottb, Rob-StLouis

Differential Revision: D19944451

fbshipit-source-id: e52a4678bf295bfc0efa4397c07f5022f30ae244