Improve SPK circuit
Summary:
Reduces lines of code in `__SPK_circuit` and provides comments explaining each line.

The main purpose for this diff is to reduce the amount of beaver triples required for each AND by taking advantage of broadcasting functionality.

Additionally, this diff removes most stacking and slicing in the circuit (reducing compute somewhat).

Reviewed By: marksibrahim

Differential Revision: D20096416

fbshipit-source-id: e5fc5e52a567dd83792244a3f2ee65e446b84b76