Add MPCTensor.from_shares
Summary:
This diff adds a static `MPCTensor.from_shares` function that allows for the creation of an `MPCTensor` from private shares. This is a first step towards adding a `crypten.nn.Module.from_shares` function, which I will add in a subsequent diff.

The diff also adds the `MPCTensor.reveal` function that was apparently missing.

Differential Revision: D20184603

fbshipit-source-id: 05bdf837acf9abead5b8b145061683012bd6070f