Fix autograd behavior when mixing with Torch tensors (#70)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/70

This diff fixes issues related to the mixing of PyTorch tensors and CrypTen tensors in (un)encrypted `crypten.nn` models. Specifically, it ensure the following behavior:

- Unencrypted model, PyTorch input: Forward, backward, parameter update work.
- Unencrypted model, CrypTen input: Prohibited.
- Encrypted model, PyTorch input: Forward, backward, parameter update work. Input is silently encrypted.
- Encrypted model, CrypTen input: Forward, backward, parameter update work.

It would in theory be possible to support the second type but it is not trivial to do: silently encrypting the entire model seems undesired, and not encrypting it is problematic because it may lead function calls to be dispatched to PyTorch rather than to CrypTen. That’s why the second type is now explicitly prohibited.

An important side effect of the diff is that it resolves the following failure case that currently exists in CrypTen: if the model is not encrypted and input is, then the parameter gradient will be `None`, which  makes `module.update_parameters()` into a no-op without warning the user.

Reviewed By: marksibrahim

Differential Revision: D20182274

fbshipit-source-id: 2bcd5991cc60172e159a12ad71a94aec0a72a5c4