Add MPCTensor.from_shares (#71)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/71

This diff adds a static `MPCTensor.from_shares` function that allows for the creation of an `MPCTensor` from private shares. This is a first step towards adding a `crypten.nn.Module.from_shares` function, which I will add in a subsequent diff.

The diff also adds the `MPCTensor.reveal` function that was apparently missing.

Reviewed By: shubho

Differential Revision: D20184603

fbshipit-source-id: b3c8b2955747f01fe665dd3b47533ef7a2aba604