Add state_dict logic to crypten.nn.Module
Summary:
This diff implements `state_dict()` and `load_state_dict()` functions in `crypten.nn.Module` that work in the same way as their PyTorch counterparts. This will make serialization of models easier: as long as `CrypTensor`s are serializable, the `state_dict` will be a serializable object that can be pickled directly.

This diff also adds the `CrypTensor.copy_()` function, which was still missing from the `CrypTensor` API.

Differential Revision: D20603904

fbshipit-source-id: bce3570a832c59293a8bbaf836c0e80a465791bb