Add state_dict logic to crypten.nn.Module (#89)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/89

This diff implements `state_dict()` and `load_state_dict()` functions in `crypten.nn.Module` that work in the same way as their PyTorch counterparts. This will make serialization of models easier: as long as `CrypTensor`s are serializable, the `state_dict` will be a serializable object that can be pickled directly.

This diff also adds the `CrypTensor.copy_()` function, which was still missing from the `CrypTensor` API.

Reviewed By: knottb

Differential Revision: D20603904

fbshipit-source-id: d88b0a59da11d7b2ce8f2ff2a16bba76a09d2863