log-reduction algorithm for argmax/max. (#189)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/189

This diff implements log-reduction algorithm for argmax/max. It refactors some of the existing argmax/max functionality in order to (1) allow for a choice of algorithms and (2) avoid re-computing the (arg)maxes when not needed. It also refactors the argmax/max tests in test_mpc so that each max algorithm is run as a separate test and therefore doesn't time out.

Reviewed By: lvdmaaten

Differential Revision: D20885416

fbshipit-source-id: c68d2fd7e18b5e9d50cd4d45936ee6f329d9b393