Double log reduction algorithm for computing max/argmax (#190)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/190

This diff implements the double log reduction algorithm for computing the argmax/max. This algorithm is suboptimal to the log reduction and the pairwise algorithms in the number of comparisons and number of communication rounds respectively, however, it is a key component of the accelerating cascading algorithm (which will be implemented next). The algorithm is described in Section 2.6.2 of https://folk.idi.ntnu.no/mlh/algkon/jaja.pdf

Also refactored a little of the log reduction implementation from the previous diff for use in the double log reduction algorithm

Reviewed By: lvdmaaten

Differential Revision: D20911043

fbshipit-source-id: db6a80db19eb7d7bf492b3d6814d3daa2cf4da72