Fix flaky max/min argmax/argmin tests (#196)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/196

The max/min and argmax/argmin tests were failing around 1 in 1000 runs due to possibly quantization. This diff changes the random tensors to be integers (that are then converted to floats) so as to  ensure the test numbers are well-separated from each other and thus avoid quantization effects. Testing on 5000 runs showed 0 failures after this fix.

Reviewed By: marksibrahim

Differential Revision: D20971734

fbshipit-source-id: df9f35ed390d6e3737526deeee59710ca2f44368