Add load/save closure functions (#201)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/201

This diff adds support for using custom load and save functions in `crypten.load` and `crypten.save`. The custom load/save functions need to have the same interface as `torch.load/torch.save`.

Note that this does not yet change the name of the function to `load_from_party` and `save_from_party` as we discussed in the design document. I plan to do that in the next diff.

Reviewed By: knottb

Differential Revision: D21025272

fbshipit-source-id: 8ae99e7b4ef0aa0ecfe99e84f4b23417f355bdaf