Add masking to update_parameters
Summary:
Added a masking kwarg to `update_parameters` in crypten.nn.Module

The mask allows for a one-time overwrite of the learning rate. This is useful to zero out the gradients for dummy inputs. (e.g. if we do a private match that returns unmatched rows, we can zero out the gradient updates for the unmatched rows by using an encrypted mask).

Differential Revision: D21723160

fbshipit-source-id: 8db203f4f4f0d0a7eeab196220c9af834665376e