Implement safe deserialization
Summary:
Python's pickle package is susceptible to RCE attacks in deserialization (using `pickle.loads()`).

This diff implements a restriced Unpickler that checks input classes for loaded objects / functions, and rejects them if they are not within a "safe set".

Also allows users to register new safe types that can be sent as objects.

Also updated objects send / recv / broadcast tests to communicate several types of safe standard / custom objects, and tests that invalid objects are rejected.

This is a 2nd attempt at D21574976, which I plan to abandon since it does not allow torch tensors to be communicated.

Differential Revision: D21745034

fbshipit-source-id: 9e7e38cbd8a2d5059cf04d1a945a0efd05a8bf58