Implement safe deserialization (#128)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/128

Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/208

Python's pickle package is susceptible to RCE attacks in deserialization (using `pickle.loads()`).

This diff implements a restriced Unpickler that checks input classes for loaded objects / functions, and rejects them if they are not within a "safe set".

Also allows users to register new safe types that can be sent as objects.

Also updated objects send / recv / broadcast tests to communicate several types of safe standard / custom objects, and tests that invalid objects are rejected.

This is a 2nd attempt at D21574976, which I plan to abandon since it does not allow torch tensors to be communicated.

Reviewed By: Cictrone

Differential Revision: D21745034

fbshipit-source-id: 560c69a14e05c8e126b0e5fbac77bb80210e296b