Modify imported onnx graphs to prevent encryption of shapes (#121)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/121

Modifies conversion of reshape modules from onnx to prevent encryption of shapes:

* Adds `__repr__` to modules
* Prevents encryption of shapes in `module.encrypt()`
* Modifies CrypTen model graph to encode shapes as `torch.Size`
* Refactors onnx conversion into a `FromOnnx` class for shared state
* Adds new tests in `test_onnx_converter` for reshape

Note architectures with variable shapes (for example `                out = out.view(out.size(0), -1)`) are not yet supported, since shapes are split across multiple constant modules in different parts of the graph.

Also note indices used for Gather are still silently decrypted. A fix for this is to come in a subsequent diff.

Reviewed By: lvdmaaten

Differential Revision: D21664942

fbshipit-source-id: 14f2d837db4963798886fcad634219915b9d6cb2