Fix batchnorm input gradient
Summary:
The gradients w.r.t. the input are computed incorrectly when batchnorm is in training mode. Specifically, the `backward()` implementation ignores the fact that the mean and variance are a function of the input, and so they contribute terms to the gradient of the input themselves.

The reason that this went unnoticed is because the batchnorm unit tests were incorrect as well: [this code](https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/deeplearning/projects/crypten/test/test_gradients.py?lines=734) does not pass the `training` input argument the way it should.

This diff fixes both the gradient and the unit test.

Differential Revision: D22275315

fbshipit-source-id: efb04944af4712f1ce58ab52bc1cad18aa869cc7