Fix batchnorm input gradient (#143)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/143

The gradients w.r.t. the input are computed incorrectly when batchnorm is in training mode. Specifically, the `backward()` implementation ignores the fact that the mean and variance are a function of the input, and so they contribute terms to the gradient of the input themselves.

The reason that this went unnoticed is because the batchnorm unit tests were incorrect as well: [this code](https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/deeplearning/projects/crypten/test/test_gradients.py?lines=734) does not pass the `training` input argument the way it should.

This diff fixes both the gradient and the unit test.

Reviewed By: marksibrahim

Differential Revision: D22275315

fbshipit-source-id: cb1cb6b19a063951ec1a95f8374171c4d80c8031