Restrict autograd backward in accordance with PyTorch
Summary:
Calling `loss.backward()` on a PyTorch tensor will only work without a `grad_input` argument when `loss` is a 1-element tensor. Otherwise it throws an error.

This diff raises the same error for cases where PyTorch would raise the error and handles cases in testing where this was not handled properly.

Differential Revision: D22309803

fbshipit-source-id: cae9e57f097ec3da5bb144d834cff738a6ae2b5f