Initialize two sets of generators (for CPU and GPU) at crypten.init()
Summary:
`crypten.init()` takes in a device arguments that specify the device for RNG generators `comm.get().g0` and `comm.get().g1`. This design is not flexible since users can not run their models on CPU if they specify CUDA as device when they call `init()`.

This diff refactor RNG generators to prepare two sets of generators during init, one for CPU and one for GPU. When user run their model on CPU, then generators on CPU will be used, and vice versa.

Differential Revision: D22362004

fbshipit-source-id: 36c11e7cb531ab5afa685147ad53fc45e8b6256a