Restrict autograd backward in accordance with PyTorch (#214)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/214

Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/146

Calling `loss.backward()` on a PyTorch tensor will only work without a `grad_input` argument when `loss` is a 1-element tensor. Otherwise it throws an error.

This diff raises the same error for cases where PyTorch would raise the error and handles cases in testing where this was not handled properly.

Reviewed By: lvdmaaten

Differential Revision: D22309803

fbshipit-source-id: 59cbe9ea486be766a3d0e33afa348b5e3cf54999