Add missing grad_output to BCEWithLogits gradient
Summary:
In `BCEWithLogitsLoss`'s gradient, we did not multiply by `grad_output`. This is usually not an issue because the `grad_output` of a loss function is typically equal to 1. However, for multi-objective losses where the output of `BCEWithLogitsLoss` is one component of the loss, this may not be the case.

This diff adds code to multiply the gradient in `BCEWithLogitsLoss` by `grad_output` as expected.

Reviewed By: lvdmaaten

Differential Revision: D22531944

fbshipit-source-id: 1fee19711775a58259d1b6cce6f32faf06e17bf6