Disable tracking of computations on CrypTensors for which requires_grad is False (#156)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/156

When `CrypTensor.requires_grad` is `False`, we are currently still storing some information: namely, we are storing a nearly empty `res` object in `self.parents`. This leads to an accumulation of data that is not removed because the user never calls `backward()` (see D21644543 and P131496553).

This diff makes sure no information is stored during forward computations with `CrypTensor`s on which `requires_grad` is `False`.

Note that this does lead to behavior change: it is no longer possible to change the `requires_grad` value to `True` *after* a forward computation has been performed to still obtain the gradient on that tensor. This behavior appears to be in line with the behavior of the PyTorch autograd, which does not seem to be taking forward computations on tensors unless `requires_grad` is `True` either:
```
>>> a = torch.randn(5)
>>> b = a.exp().sum()
>>> a.requires_grad = True
>>> b.requires_grad = True
>>> b.backward()
>>> a.grad
>>> print(a.grad)
None
>>> b.grad
tensor(1.)
```

The diff also fixes a small issue where running `t.backward()` on a `CrypTensor` of size 1 with `requires_grad=True` did not set `t.grad` to 1 (akin to the PyTorch example above). I have added a unit test that locks in these behavior changes.

Reviewed By: knottb

Differential Revision: D22509857

fbshipit-source-id: 90e9143cf548df866172993c2b1855293b56551f