Fix the API of MPCTensor.to
Summary:
`MPCTensor.to` has a required keyword `input`. If the user want to call `MPCTensor.to(device="cuda")`, then it would throw an error saying that `input` is required.

The desired behavior of `to` is that if user provides an arg with type `Ptype` or a kwargs called `ptype`, then the function will call `to_ptype`. Otherwise it will call `torch.to` to convert underlying tensor to the corresponding `dtype` or `device`.

This diff modifies the function to the desired behavior.

Reviewed By: knottb

Differential Revision: D22664985

fbshipit-source-id: 96c46bab4bfed5f1f5c6b371d2c1c3e90d51a72b