Add SGD optimizer
Summary:
Added `crypten.optim` with `SGD` optimizer that can update parameters using `step()` as in PyTorch adapted from `torch.optim` to remove requirements that parameters be `torch.Tensors` and replacing some of the type checks that break.

Also added a test in `test_optim` that ensures that sgd works as expected.

Differential Revision: D22850046

fbshipit-source-id: 6521dd0bdbe9694a5120c3c415f7a1c509e4d25e