Parameterize CUDALongTensor's cross multiplication (#165)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/165

CUDALongTensor utilizes the cross multiplication trick to embed computation over LongTensor into FloatTensor. In D22288266 (https://github.com/facebookresearch/CrypTen/commit/64eb7b20d12005cd4141d814a17e1864fe83bd33) we made an optimization to this trick by reducing the number of block to 3. It is shown on the benchmark that this optimization would result in a blowup in error for `mul` and `matmul`. Previous construction with 4 blocks do not have this issue.

This diff revert the construction to the previous iteration.

Also, parameterize the cross-multiplication technique.

Reviewed By: knottb

Differential Revision: D22933262

fbshipit-source-id: bda28b5940c34870b66f683e02ac6870e340d384