Add SGD optimizer (#217)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/217

Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/164

Added `crypten.optim` with `SGD` optimizer that can update parameters using `step()` as in PyTorch adapted from `torch.optim` to remove requirements that parameters be `torch.Tensors` and replacing some of the type checks that break.

Also added a test in `test_optim` that ensures that sgd works as expected.

Differential Revision: D22850046

fbshipit-source-id: 91ffc6d78fee2fb633a13ec53051a0e15452e8a0