Fix autograd issue when node has multiple parents
Summary:
When a node in the backwards pass has multiple parents, the current autograd implementation does not accumulate the gradients coming from those parents. This diff addresses that issue.

Curiously, the unit tests had tests for this issue, but for some reason they did not fire. I have confirmed that the tests run, and added some additional test cases.

Differential Revision: D23287670

fbshipit-source-id: 16ad2b51b4139634d1bcd6d464d96bd14911e203