Support calling torch.{cat,stack} on CrypTensors
Summary: At present, it is somewhat tricky to write code that works both with PyTorch and with CrypTen tensors because static functions like `torch.cat` do not work on CrypTen tensors (and vice versa). This diff addresses this problem by allowing users to call `torch.{cat,stack}(list_of_cryptensors)`. Calls are dispatched to CrypTen and go through the CrypTen autograd as expected.

Differential Revision: D23668830

fbshipit-source-id: d0f324e0fdb051fd6a9842f6591e3244621d86e4