Support calling torch.{cat,stack} on CrypTensors (#176)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/176

At present, it is somewhat tricky to write code that works both with PyTorch and with CrypTen tensors because static functions like `torch.cat` do not work on CrypTen tensors (and vice versa). This diff addresses this problem by allowing users to call `torch.{cat,stack}(list_of_cryptensors)`. Calls are dispatched to CrypTen and go through the CrypTen autograd as expected.

[The `__torch_function__` functionality was added in PyTorch 1.5](https://github.com/pytorch/pytorch/releases?after=v1.5.1-rc1) so this diff also bumps the PyTorch dependency to 1.5.

Reviewed By: knottb

Differential Revision: D23668830

fbshipit-source-id: 663a5faffd42356338f63b0b83bed517b35e4bc2