Remove dim from kwargs in max function (#223)
Summary:
Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/223

This diff ensures that `dim` is removed from `kwargs` in the max forward function. This is necessary since `dim `is passed in separately to the max function at the end of function, and when `dim` is present in `kwargs`, this results in an error. Note that `dim` is already extracted from `kwargs` if necessary.

Reviewed By: lvdmaaten

Differential Revision: D23773999

fbshipit-source-id: 2de45261a16fdb29f84533dc51d086cb1a9ec9f1