Remove silent encryption / decryption behavior
Summary:
At present, `crypten.nn` sometimes silently does not encrypt values (in `Constant`), it silently encrypts values (when a PyTorch tensors is used as input into an encrypted model), and silently decrypts indices in the `nn.Gather` module. Such behavior is dangerous: in particular, in models that have an embedding as first layer it may inadvertently decrypt the data.

This diff removes all silent encryption / decryption behavior, or adds warnings when some value is not encrypted. Note that these changes are breaking changes.

Differential Revision: D27474350

fbshipit-source-id: 8f2ec72233ad77f5451472171e9908651a51191e