Fix groups argument to convolution on GPU
Summary:
In `gradients.py`, we [pass the `groups` argument to the underlying kernel](https://fburl.com/diffusion/3egy7m4m). This works correctly on CPU, but fails on GPU because there we use the `groups` argument to represent long integers as a group of floats.

This diff addresses that issue.

Differential Revision: D28385620

fbshipit-source-id: 6ef9a10d80b54ffacc73e38d0844d4ea69c4bdd9