Fix groups argument to convolution on GPU (#260)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/260

In `gradients.py`, we [pass the `groups` argument to the underlying kernel](https://fburl.com/diffusion/3egy7m4m). This works correctly on CPU, but fails on GPU because there we use the `groups` argument to represent long integers as a group of floats.

This diff addresses that issue.

Reviewed By: knottb

Differential Revision: D28385620

fbshipit-source-id: 354d53526ea7356e9d6ff2bae141b3292ed7eded