Make value in Constant not be a buffer
Summary:
At present, `value` in `Constant{OfShape}` is a buffer. This means it gets copied to GPU when a model gets copied. But, in practice, we only use it to concatenate values to a `Shape`, which is not a buffer and always remains on CPU (as it should). This leads to problems, for example, in vision transformers we end up concatenating a CPU tensor (from `Shape`) with a GPU tensor (from `Constant`).

I don't know what the right solution here is long-term, but this seems to work for fine in practice.

Differential Revision: D28444998

fbshipit-source-id: b07d01078682721da06aee3f28ea5540cd890c0e