Make value in Constant not be a buffer (#262)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/262

At present, `value` in `Constant{OfShape}` is a buffer. This means it gets copied to GPU when a model gets copied. But, in practice, we only use it to concatenate values to a `Shape`, which is not a buffer and always remains on CPU (as it should). This leads to problems, for example, in vision transformers we end up concatenating a CPU tensor (from `Shape`) with a GPU tensor (from `Constant`).

I don't know what the right solution here is long-term, but this seems to work for fine in practice.

Reviewed By: knottb

Differential Revision: D28444998

fbshipit-source-id: e93ec87e951c7c4e9b32d89a4a1adb36657faeec