Fix initializations of tensors to the GPU.
Closes #250.

CUDALongTensors are now initialized to be on the same device as the
input "data" tensor if (a) no device is explicitly specified and (b) the
"data" tensor is already on a GPU. CrypTen previously would default to
putting the tensor on cuda:0, which could raise problems in cases where
the user wants to use multiple GPUs.

This commit also fixes some locations where a tensor would be created on
CPU rather than GPU, which could also lead to errors where we try to
perform operations on two tensors that are on different devices.