Clean up values in nn.Graph that are no longer needed
Summary:
`crypten.nn.Graph` currently maintains references to all intermediate values that it has computed, even if those values are no longer needed. This may lead to memory inefficiencies, in particular, when models are run in inference mode / autograd is disabled (in which case autograd does not maintain references to intermediate results).

The diff adds logic to `crypten.nn.Graph.forward()` that checks whether or not we still need intermediate values for future computations, and removes references to values that are no longer needed.

Reviewed By: knottb

Differential Revision: D28570323

fbshipit-source-id: 27d2ff67487e35d7ff5895f3cb37326d180312fa