Make PRSS more memory-efficient
Summary: The current PRSS implementation creates a list of all tensors for all parties before performing the reduction via (arithmetic or binary) addition. This leads to memory issues when the number of parties involved in the computation is large and the tensors live on GPU (in which case all the tensors live on the same GPU). This diff replaces the existing implementation by an implementations that samples tensors one-by-one and performs the reduction on-the-fly.

Differential Revision: D28650678

fbshipit-source-id: dfdb27aaa2b27b256122b43792f8645112381a31