Fix incorrect calls into MPI collectives
Summary:
There are several call-sites in CrypTen where MPI collectives are called with keyword arguments in the function call, whereas the function expects an unnamed argument. Python allows this so it generally works, but in combination with the changes in D28698384 it leads to problems because that relies on the value of `args` to log the communication -- it cannot find those values if they are erroneously in `kwargs`.

This diff corrects the incorrect function calls.

Reviewed By: knottb

Differential Revision: D28721879

fbshipit-source-id: 46d06783fbfffc1d8e1065441780d0d451e7beb2