Fix bug in integer division
Summary:
In certain cases, when inputs are integers, division will statistically round rather than converting to floating point division. This happens when the encoder scale is set to 1, which will happen after a comparison operator (e.g. sign() etc.)

I found this in the linear_svm example during evaluation as we compute `accuracy = correct.add(1).div(2).mean()`. The `div(2)` will round because the input encoder has scale 1 due to the sign in the SVM code.

This diff corrects the issue and adds this edge case to testing.

Differential Revision: D33686267

fbshipit-source-id: 06da5d3986120d8bf64b2db438586fcf3e4d9a6f